Added reasoning output to vLLM backend; moved HF backed to separate file; added reasoning level to vLLM backend for oss
