+++ charge/clients/Client.py
@@ -114,3 +114,11 @@ class Client:
         parser.add_argument(
             "--server-urls", nargs="*", type=str, default="http://127.0.0.1:8000/sse"
         )
+        parser.add_argument(
+            "--reasoning-effort",
+            type=str,
+            choices=["low", "medium", "high"],
+            default="medium",
+            help="Effort level for reasoning models",
+        )
+
